kestrel-kernels/pyproject.toml
[build-system]
requires = [
  "scikit-build-core>=0.10",
  "cmake>=3.26",
  "ninja",
  "torch==2.9.1",
  "nvidia-cutlass-dsl==4.3.3",
  "apache-tvm-ffi",
  "torch-c-dlpack-ext",
  "wheel",
]
build-backend = "scikit_build_core.build"

[project]
name = "kestrel-kernels"
version = "0.1.0"
description = "CUDA kernel library for Kestrel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "nvidia-cutlass-dsl==4.3.3",
  "apache-tvm-ffi",
  "torch-c-dlpack-ext",
  "torch==2.9.1",
]

[tool.scikit-build]
cmake.build-type = "Release"
minimum-version = "build-system.requires"
wheel.packages = []  # Disable auto-include - CMake install() controls contents
cmake.args = ["-DCMAKE_CUDA_ARCHITECTURES=90a"]
